If minChars===0 then show all results immediately
This is a port of a patch made from https://github.com/coderow/jquery-tokeninput (a year ago) which has not been merged with the mainline repo for this project.  The patch ported is 29bb94f6f8c3739992c285be9e645e7f2301d74c.
Updated to use spaces for indentation to be consistent with the rest of the file

See:
https://github.com/coderow/jquery-tokeninput/commit/29bb94f6f8c3739992c285be9e645e7f2301d74c